Better organize ValueFunctionPlayer, AlphaBeta and tree search utils
