+++ catanatron_experimental/catanatron_experimental/machine_learning/players/mcts.py
@@ -9,8 +9,6 @@ from catanatron.models.player import Player
 from catanatron_experimental.machine_learning.players.playouts import run_playout
 from catanatron_experimental.machine_learning.players.tree_search_utils import (
     execute_spectrum,
-)
-from catanatron_experimental.machine_learning.players.minimax import (
     list_prunned_actions,
 )
