+++ catanatron_experimental/catanatron_experimental/machine_learning/players/tree_search_utils.py
@@ -8,9 +8,17 @@ from catanatron.models.decks import (
 from catanatron.models.enums import (
     DEVELOPMENT_CARDS,
     RESOURCES,
+    SETTLEMENT,
+    CITY,
     Action,
     ActionType,
 )
+
+from catanatron.state_functions import get_player_buildings
+from catanatron_gym.features import (
+    build_production_features,
+)
+from catanatron_experimental.machine_learning.players.value import value_production
 
 
 def execute_spectrum(game, action):
@@ -93,3 +101,79 @@ def expand_spectrum(game, actions):
         outprobas = execute_spectrum(game, action)
         children[action] = outprobas
     return children  # action => (game, proba)[]
+
+
+def list_prunned_actions(game):
+    current_color = game.state.current_color()
+    playable_actions = game.state.playable_actions
+    actions = playable_actions.copy()
+    types = set(map(lambda a: a.action_type, playable_actions))
+
+    # Prune Initial Settlements at 1-tile places
+    if ActionType.BUILD_SETTLEMENT in types and game.state.is_initial_build_phase:
+        actions = filter(
+            lambda a: len(game.state.board.map.adjacent_tiles[a.value]) != 1, actions
+        )
+
+    # Prune Trading if can hold for resources. Only for rare resources.
+    if ActionType.MARITIME_TRADE in types:
+        port_resources = game.state.board.get_player_port_resources(current_color)
+        has_three_to_one = None in port_resources
+        # TODO: for 2:1 ports, skip any 3:1 or 4:1 trades
+        # TODO: if can_safely_hold, prune all
+        tmp_actions = []
+        for action in actions:
+            if action.action_type != ActionType.MARITIME_TRADE:
+                tmp_actions.append(action)
+                continue
+            # has 3:1, skip any 4:1 trades
+            if has_three_to_one and action.value[3] is not None:
+                continue
+            tmp_actions.append(action)
+        actions = tmp_actions
+
+    if ActionType.MOVE_ROBBER in types:
+        actions = prune_robber_actions(current_color, game, actions)
+
+    return list(actions)
+
+
+def prune_robber_actions(current_color, game, actions):
+    """Eliminate all but the most impactful tile"""
+    enemy_color = next(filter(lambda c: c != current_color, game.state.colors))
+    enemy_owned_tiles = set()
+    for node_id in get_player_buildings(game.state, enemy_color, SETTLEMENT):
+        enemy_owned_tiles.update(game.state.board.map.adjacent_tiles[node_id])
+    for node_id in get_player_buildings(game.state, enemy_color, CITY):
+        enemy_owned_tiles.update(game.state.board.map.adjacent_tiles[node_id])
+
+    robber_moves = set(
+        filter(
+            lambda a: a.action_type == ActionType.MOVE_ROBBER
+            and game.state.board.map.tiles[a.value[0]] in enemy_owned_tiles,
+            actions,
+        )
+    )
+
+    production_features = build_production_features(True)
+
+    def impact(action):
+        game_copy = game.copy()
+        game_copy.execute(action)
+
+        our_production_sample = production_features(game_copy, current_color)
+        enemy_production_sample = production_features(game_copy, current_color)
+        production = value_production(our_production_sample, "P0")
+        enemy_production = value_production(enemy_production_sample, "P1")
+
+        return enemy_production - production
+
+    most_impactful_robber_action = max(
+        robber_moves, key=impact
+    )  # most production and variety producing
+    actions = filter(
+        # lambda a: a.action_type != action_type or a == most_impactful_robber_action,
+        lambda a: a.action_type != ActionType.MOVE_ROBBER or a in robber_moves,
+        actions,
+    )
+    return actions
