+++ catanatron_experimental/catanatron_experimental/machine_learning/players/value.py
@@ -1,0 +1,194 @@
+import random
+
+from catanatron.state_functions import (
+    get_longest_road_length,
+    get_played_dev_cards,
+    player_key,
+    player_num_dev_cards,
+    player_num_resource_cards,
+)
+from catanatron.models.player import Player
+from catanatron.models.enums import RESOURCES, SETTLEMENT, CITY
+from catanatron_gym.features import (
+    build_production_features,
+    reachability_features,
+    resource_hand_features,
+)
+
+TRANSLATE_VARIETY = 4  # i.e. each new resource is like 4 production points
+
+DEFAULT_WEIGHTS = {
+    # Where to place. Note winning is best at all costs
+    "public_vps": 3e14,
+    "production": 1e8,
+    "enemy_production": -1e8,
+    "num_tiles": 1,
+    # Towards where to expand and when
+    "reachable_production_0": 0,
+    "reachable_production_1": 1e4,
+    "buildable_nodes": 1e3,
+    "longest_road": 10,
+    # Hand, when to hold and when to use.
+    "hand_synergy": 1e2,
+    "hand_resources": 1,
+    "discard_penalty": -5,
+    "hand_devs": 10,
+    "army_size": 10.1,
+}
+
+# Change these to play around with new values
+CONTENDER_WEIGHTS = {
+    "public_vps": 300000000000001.94,
+    "production": 100000002.04188395,
+    "enemy_production": -99999998.03389844,
+    "num_tiles": 2.91440418,
+    "reachable_production_0": 2.03820085,
+    "reachable_production_1": 10002.018773150001,
+    "buildable_nodes": 1001.86278466,
+    "longest_road": 12.127388499999999,
+    "hand_synergy": 102.40606877,
+    "hand_resources": 2.43644327,
+    "discard_penalty": -3.00141993,
+    "hand_devs": 10.721669799999999,
+    "army_size": 12.93844622,
+}
+
+
+def base_fn(params=DEFAULT_WEIGHTS):
+    def fn(game, p0_color):
+        production_features = build_production_features(True)
+        our_production_sample = production_features(game, p0_color)
+        enemy_production_sample = production_features(game, p0_color)
+        production = value_production(our_production_sample, "P0")
+        enemy_production = value_production(enemy_production_sample, "P1", False)
+
+        key = player_key(game.state, p0_color)
+        longest_road_length = get_longest_road_length(game.state, p0_color)
+
+        reachability_sample = reachability_features(game, p0_color, 2)
+        features = [f"P0_0_ROAD_REACHABLE_{resource}" for resource in RESOURCES]
+        reachable_production_at_zero = sum([reachability_sample[f] for f in features])
+        features = [f"P0_1_ROAD_REACHABLE_{resource}" for resource in RESOURCES]
+        reachable_production_at_one = sum([reachability_sample[f] for f in features])
+
+        hand_sample = resource_hand_features(game, p0_color)
+        features = [f"P0_{resource}_IN_HAND" for resource in RESOURCES]
+        distance_to_city = (
+            max(2 - hand_sample["P0_WHEAT_IN_HAND"], 0)
+            + max(3 - hand_sample["P0_ORE_IN_HAND"], 0)
+        ) / 5.0  # 0 means good. 1 means bad.
+        distance_to_settlement = (
+            max(1 - hand_sample["P0_WHEAT_IN_HAND"], 0)
+            + max(1 - hand_sample["P0_SHEEP_IN_HAND"], 0)
+            + max(1 - hand_sample["P0_BRICK_IN_HAND"], 0)
+            + max(1 - hand_sample["P0_WOOD_IN_HAND"], 0)
+        ) / 4.0  # 0 means good. 1 means bad.
+        hand_synergy = (2 - distance_to_city - distance_to_settlement) / 2
+
+        num_in_hand = player_num_resource_cards(game.state, p0_color)
+        discard_penalty = params["discard_penalty"] if num_in_hand > 7 else 0
+
+        # blockability
+        buildings = game.state.buildings_by_color[p0_color]
+        owned_nodes = buildings[SETTLEMENT] + buildings[CITY]
+        owned_tiles = set()
+        for n in owned_nodes:
+            owned_tiles.update(game.state.board.map.adjacent_tiles[n])
+        num_tiles = len(owned_tiles)
+
+        # TODO: Simplify to linear(?)
+        num_buildable_nodes = len(game.state.board.buildable_node_ids(p0_color))
+        longest_road_factor = (
+            params["longest_road"] if num_buildable_nodes == 0 else 0.1
+        )
+
+        return float(
+            game.state.player_state[f"{key}_VICTORY_POINTS"] * params["public_vps"]
+            + production * params["production"]
+            + enemy_production * params["enemy_production"]
+            + reachable_production_at_zero * params["reachable_production_0"]
+            + reachable_production_at_one * params["reachable_production_1"]
+            + hand_synergy * params["hand_synergy"]
+            + num_buildable_nodes * params["buildable_nodes"]
+            + num_tiles * params["num_tiles"]
+            + num_in_hand * params["hand_resources"]
+            + discard_penalty
+            + longest_road_length * longest_road_factor
+            + player_num_dev_cards(game.state, p0_color) * params["hand_devs"]
+            + get_played_dev_cards(game.state, p0_color, "KNIGHT") * params["army_size"]
+        )
+
+    return fn
+
+
+def value_production(sample, player_name="P0", include_variety=True):
+    proba_point = 2.778 / 100
+    features = [
+        f"EFFECTIVE_{player_name}_WHEAT_PRODUCTION",
+        f"EFFECTIVE_{player_name}_ORE_PRODUCTION",
+        f"EFFECTIVE_{player_name}_SHEEP_PRODUCTION",
+        f"EFFECTIVE_{player_name}_WOOD_PRODUCTION",
+        f"EFFECTIVE_{player_name}_BRICK_PRODUCTION",
+    ]
+    prod_sum = sum([sample[f] for f in features])
+    prod_variety = (
+        sum([sample[f] != 0 for f in features]) * TRANSLATE_VARIETY * proba_point
+    )
+    return prod_sum + (0 if not include_variety else prod_variety)
+
+
+def contender_fn(params):
+    return base_fn(params or CONTENDER_WEIGHTS)
+
+
+class ValueFunctionPlayer(Player):
+    """
+    Player that selects the move that maximizes a heuristic value function.
+
+    For now, the base value function only considers 1 enemy player.
+    """
+
+    def __init__(
+        self, color, value_fn_builder_name=None, params=None, is_bot=True, epsilon=None
+    ):
+        super().__init__(color, is_bot)
+        self.value_fn_builder_name = (
+            "contender_fn" if value_fn_builder_name == "C" else "base_fn"
+        )
+        self.params = params
+        self.epsilon = epsilon
+
+    def decide(self, game, playable_actions):
+        if len(playable_actions) == 1:
+            return playable_actions[0]
+
+        if self.epsilon is not None and random.random() < self.epsilon:
+            return random.choice(playable_actions)
+
+        best_value = float("-inf")
+        best_action = None
+        for action in playable_actions:
+            game_copy = game.copy()
+            game_copy.execute(action)
+
+            value_fn = get_value_fn(self.value_fn_builder_name, self.params)
+            value = value_fn(game_copy, self.color)
+            if value > best_value:
+                best_value = value
+                best_action = action
+
+        return best_action
+
+    def __str__(self):
+        return super().__str__() + f"(value_fn={self.value_fn_builder_name})"
+
+
+def get_value_fn(name, params, value_function=None):
+    if value_function is not None:
+        return value_function
+    elif name == "base_fn":
+        return base_fn(DEFAULT_WEIGHTS)
+    elif name == "contender_fn":
+        return contender_fn(params)
+    else:
+        raise ValueError
